Show alert when status is not normal
